# tests/golden/contracts/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Greenhouse API",
  "description": "Smart plant irrigation system with evidence-based plant care, multi-sensor conflict resolution, and self-learning irrigation profiles.\n\n",
  "version": "1.0.0"
 },
 "paths": {
...
  "/api/v1/clusters": {
   "get": {
    "tags": [
     "clusters"
    ],
    "summary": "List all clusters",
    "description": "List every cluster in the system.\n\nReturns:\n    All clusters, ordered by name.",
    "operationId": "list_clusters",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/ClusterResponse"
         },
         "type": "array",
         "title": "Response List Clusters"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   },
   "post": {
    "tags": [
     "clusters"
    ],
    "summary": "Create a cluster",
    "description": "Create a new plant cluster.\n\nA cluster groups plants that share an irrigator and are watered together;\nirrigation decisions are made per-cluster, driven by the driest plant.\n\nArgs:\n    request: Cluster name, optional location label, and `environment`\n        (`indoor` or `outdoor`; affects how temperature is resolved).\n\nReturns:\n    The newly created cluster including its assigned ID.",
    "operationId": "create_cluster",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/CreateClusterRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ClusterResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/v1/clusters/{cluster_id}": {
   "get": {
    "tags": [
     "clusters"
    ],
    "summary": "Get a cluster by ID",
    "description": "Fetch a single cluster by ID.\n\nArgs:\n    cluster_id: Numeric cluster identifier.\n\nReturns:\n    The cluster row. Use ``GET /clusters/{cluster_id}/detail`` to inline\n    the cluster's plants, sensors, irrigators, config, and windows in a\n    single round-trip.\n\nRaises:\n    HTTPException: 404 if no cluster with that ID exists.",
    "operationId": "get_cluster",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "cluster_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Cluster Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ClusterResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "put": {
    "tags": [
     "clusters"
    ],
    "summary": "Update a cluster",
    "description": "Partially update a cluster's metadata.\n\nOnly fields present in the request body are modified; omitted fields are\nleft unchanged.\n\nArgs:\n    cluster_id: Numeric cluster identifier.\n    request: Fields to update — any combination of name, location, and\n        environment.\n\nReturns:\n    The updated cluster.\n\nRaises:\n    HTTPException: 404 if no cluster with that ID exists.",
    "operationId": "update_cluster",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "cluster_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Cluster Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/UpdateClusterRequest"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ClusterResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "delete": {
    "tags": [
     "clusters"
    ],
    "summary": "Delete a cluster",
    "description": "Delete a cluster and all its associated data.\n\nCascades to plants, sensors, irrigators, and irrigation config. This\noperation is irreversible.\n\nArgs:\n    cluster_id: Numeric cluster identifier.\n\nReturns:\n    success=True on successful deletion.\n\nRaises:\n    HTTPException: 404 if no cluster with that ID exists.",
    "operationId": "delete_cluster",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "cluster_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Cluster Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SuccessResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/clusters/{cluster_id}/detail": {
   "get": {
    "tags": [
     "clusters"
    ],
    "summary": "Get a cluster with its child resources inlined",
    "description": "Return a cluster together with every child resource in one round-trip.\n\nInlines the cluster's plants, sensors, irrigators, irrigation config, and\nirrigation windows so a UI or MCP tool can render a full cluster view\nwithout issuing a fan-out of follow-up requests. The plain\n``GET /clusters/{cluster_id}`` endpoint remains unchanged for clients that\nonly need the cluster row itself.\n\nArgs:\n    cluster_id: Numeric cluster identifier.\n    expand: Reserved for future expansion variants — currently only\n        ``children`` is honored (and is also the default).\n\nReturns:\n    The cluster wrapped together with plants, sensors, irrigators,\n    irrigation config, and irrigation windows.\n\nRaises:\n    HTTPException: 404 if no cluster with that ID exists.",
    "operationId": "get_cluster_detail",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "cluster_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Cluster Id"
      }
     },
     {
      "name": "expand",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "description": "Reserved for future expansion levels — currently must be ``children`` (the default).",
       "default": "children",
       "title": "Expand"
      },
      "description": "Reserved for future expansion levels — currently must be ``children`` (the default)."
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ClusterDetailResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "ClusterDetailResponse": {
    "properties": {
     "cluster": {
      "$ref": "#/components/schemas/ClusterResponse"
     },
     "plants": {
      "items": {
       "$ref": "#/components/schemas/PlantResponse"
      },
      "type": "array",
      "title": "Plants"
     },
     "sensors": {
      "items": {
       "$ref": "#/components/schemas/SensorResponse"
      },
      "type": "array",
      "title": "Sensors"
     },
     "irrigator": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/IrrigatorResponse"
       },
       {
        "type": "null"
       }
      ]
     },
     "config": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/ConfigResponse"
       },
       {
        "type": "null"
       }
      ]
     },
     "windows": {
      "items": {
       "$ref": "#/components/schemas/IrrigationWindowResponse"
      },
      "type": "array",
      "title": "Windows"
     }
    },
    "type": "object",
    "required": [
     "cluster",
     "plants",
     "sensors",
     "irrigator",
     "config",
     "windows"
    ],
    "title": "ClusterDetailResponse",
    "description": "Cluster plus its child resources (plants, sensors, irrigators, config, windows).\n\nReturned by GET /clusters/{id}?expand=children. Plain GET without the\nexpand flag continues to return ClusterResponse for back-compat."
   },
...
   "ClusterResponse": {
    "properties": {
     "name": {
      "type": "string",
      "title": "Name"
     },
     "location": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Location"
     },
     "environment": {
      "type": "string",
      "title": "Environment",
      "default": "indoor"
     },
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "created_at": {
      "type": "integer",
      "title": "Created At"
     }
    },
    "type": "object",
    "required": [
     "name",
     "id",
     "created_at"
    ],
    "title": "ClusterResponse",
    "description": "A cluster as stored."
   },
...
   "ConfigResponse": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "cluster_id": {
      "type": "integer",
      "title": "Cluster Id"
     },
     "mode": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Mode"
     },
     "duration_minutes": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Duration Minutes"
     },
     "interval_hours": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Interval Hours"
     },
     "auto_run": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Auto Run"
     },
     "last_updated": {
      "type": "integer",
      "title": "Last Updated"
     },
     "daily_cap_minutes": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Daily Cap Minutes"
     },
     "max_events_per_day": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Max Events Per Day"
     },
     "quiet_start_hour": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Quiet Start Hour"
     },
     "quiet_end_hour": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Quiet End Hour"
     }
    },
    "type": "object",
    "required": [
     "id",
     "cluster_id",
     "mode",
     "duration_minutes",
     "interval_hours",
     "auto_run",
     "last_updated"
    ],
    "title": "ConfigResponse",
    "description": "Declared (raw) cluster config — nulls represent inherited values."
   },
   "CreateClusterRequest": {
    "properties": {
     "name": {
      "type": "string",
      "title": "Name"
     },
     "location": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Location"
     },
     "environment": {
      "type": "string",
      "title": "Environment",
      "default": "indoor"
     }
    },
    "type": "object",
    "required": [
     "name"
    ],
    "title": "CreateClusterRequest",
    "description": "Body for creating a cluster."
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "IrrigationWindowResponse": {
    "properties": {
     "start_hour": {
      "type": "integer",
      "title": "Start Hour"
     },
     "end_hour": {
      "type": "integer",
      "title": "End Hour"
     },
     "weekday_mask": {
      "type": "integer",
      "title": "Weekday Mask",
      "default": 127
     },
     "label": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Label"
     },
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "cluster_id": {
      "type": "integer",
      "title": "Cluster Id"
     }
    },
    "type": "object",
    "required": [
     "start_hour",
     "end_hour",
     "id",
     "cluster_id"
    ],
    "title": "IrrigationWindowResponse",
    "description": "A watering window as stored."
   },
...
   "IrrigatorResponse": {
    "properties": {
     "tuya_device_id": {
      "type": "string",
      "title": "Tuya Device Id"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "type": {
      "type": "string",
      "title": "Type"
     },
     "config": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Config"
     },
     "reservoir_l": {
      "anyOf": [
       {
        "type": "number",
        "minimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Reservoir L"
     },
     "flow_rate_l_per_min": {
      "anyOf": [
       {
        "type": "number",
        "minimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Flow Rate L Per Min"
     },
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "cluster_id": {
      "type": "integer",
      "title": "Cluster Id"
     }
    },
    "type": "object",
    "required": [
     "tuya_device_id",
     "name",
     "type",
     "id",
     "cluster_id"
    ],
    "title": "IrrigatorResponse",
    "description": "An irrigator as stored, with the cluster it waters."
   },
...
   "PlantResponse": {
    "properties": {
     "species": {
      "type": "string",
      "title": "Species"
     },
     "category": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Category"
     },
     "water_needs": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Water Needs"
     },
     "light_needs": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Light Needs"
     },
     "ideal_temp_min": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Ideal Temp Min"
     },
     "ideal_temp_max": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Ideal Temp Max"
     },
     "ideal_humidity_min": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Ideal Humidity Min"
     },
     "ideal_humidity_max": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Ideal Humidity Max"
     },
     "notes": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Notes"
     },
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "cluster_id": {
      "type": "integer",
      "title": "Cluster Id"
     }
    },
    "type": "object",
    "required": [
     "species",
     "id",
     "cluster_id"
    ],
    "title": "PlantResponse",
    "description": "A plant as stored, with the cluster it belongs to."
   },
...
   "SensorResponse": {
    "properties": {
     "tuya_device_id": {
      "type": "string",
      "title": "Tuya Device Id"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "type": {
      "type": "string",
      "title": "Type"
     },
     "config": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Config"
     },
     "plant_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Plant Id"
     },
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "cluster_id": {
      "type": "integer",
      "title": "Cluster Id"
     }
    },
    "type": "object",
    "required": [
     "tuya_device_id",
     "name",
     "type",
     "id",
     "cluster_id"
    ],
    "title": "SensorResponse",
    "description": "A sensor as stored, with its cluster and current plant assignment."
   },
...
   "SuccessResponse": {
    "properties": {
     "success": {
      "type": "boolean",
      "title": "Success"
     }
    },
    "type": "object",
    "required": [
     "success"
    ],
    "title": "SuccessResponse",
    "description": "Generic success acknowledgement."
   },
...
   "UpdateClusterRequest": {
    "properties": {
     "name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Name"
     },
     "location": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Location"
     },
     "environment": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Environment"
     }
    },
    "type": "object",
    "title": "UpdateClusterRequest",
    "description": "Partial update of a cluster; omitted fields keep their value."
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  },
  "securitySchemes": {
   "HTTPBearer": {
    "type": "http",
    "scheme": "bearer"
   }
  }
 },
...
}